* exceptions.cc (handle_exceptions): Just "core dump" if SIGSEGV in signal thread. * external.cc (fillout_pinfo): Fix compiler warning. * sigproc.h: Eliminate special asm naming for sig_dispatch_pending. * sigproc.cc (sig_send): Remove debugging statements.

+++ winsup/cygwin/external.cc
@@ -48,7 +48,7 @@ fillout_pinfo (pid_t pid, int winpid)
 
       if (!p)
 	{
-	  if (!winpid || (!nextpid && thispid != pid))
+	  if (!winpid || (!nextpid && thispid != (DWORD) pid))
 	    continue;
 	  ep.pid = thispid;
 	  ep.dwProcessId = cygwin_pid (thispid);
